extensions/GCS-v7-with-empathy/backend/requirements.txt
numpy>=1.23
tensorflow>=2.12
pyyaml>=6.0
cryptography>=41.0
scikit-learn>=1.2
z3-solver>=4.12.0
# Logging is built-in, but you may want loguru or similar for advanced logging
# loguru>=0.7
# If you use argparse and other stdlib modules, no need to list them
